reduced blanks data

+++ convert_mhd2jpg_png_LiTS_reduced_blanks.py
@@ -14,7 +14,7 @@ import cv2
 import numpy as np
 import scipy.misc
 
-DATA_DIRECTORY = '/home/victor/LITS'
+DATA_DIRECTORY = '/home/victor/newLITS'
 OUT_DIRECTORY = "/home/victor/newLITS_reduced_blanks"
 
 
@@ -75,10 +75,18 @@ def ndarry2jpg_png((data_file, img_gt_file, out_dir)):
 
     img_pad = np.pad(img, ((0, 0), (0, 0), (1, 1)), 'constant', constant_values=(0, 0))
 
+    num_empty_to_keep = int(math.ceil(((spacing[0] + spacing[1]) / 2.0) / 0.7 * 10))
+    assert num_empty_to_keep >= 0
+
     for i in xrange(0, num_slices):
         if i >= img_gt.shape[2]:
             print data_file, img_gt_file, spacing, gt_spacing, 'fail_idx!'
             continue
+
+        if np.array_equal(np.unique(img_gt[:, :, i - num_empty_to_keep:i + num_empty_to_keep + 1]), [0]):
+            print i, num_empty_to_keep, data_file, img_gt_file, spacing, gt_spacing, 'fail_empty!'
+            continue
+
         img3c = img_pad[:, :, i:i + 3]
         scipy.misc.imsave(os.path.join(out_dir, "JPEGImages", fn + "_" + str(i) + ".jpg"), img3c)
         cv2.imwrite(os.path.join(out_dir, "PNGImages", fn_gt + "_" + str(i) + ".png"), img_gt[:, :, i])
@@ -107,7 +115,7 @@ def convert(data_dir, out_dir):
     if not os.path.exists(os.path.join(out_dir, "dataset")):
         os.mkdir(os.path.join(out_dir, "dataset"))
 
-    p = multiprocessing.Pool(7)
+    p = multiprocessing.Pool(3)
     retval = p.map(ndarry2jpg_png, zip(vols, segs, itertools.repeat(out_dir, len(vols))))
     p.close()
